Align budget and expenses test files with review naming guidance.
Consolidate expenses integration coverage into expenses.integration.test.js, rename budget integration files for API vs lib clarity, and expand lib.budget.unit.test.js with additional malformed and valid-case checks.
